requirements.txt
aiogram==2.22
python-dotenv==0.20.0
